add basic username validation to server

+++ routes/ideas.js
@@ -41,7 +41,10 @@ router.post("/", async (req, res) => {
 // update  idea
 router.put("/:id", async (req, res) => {
   try {
-    const updatedIdea = await Idea.findByIdAndUpdate(
+
+    const idea = await Idea.findById(req.params.id)
+    if(idea.username === req.body.username){
+      const updatedIdea = await Idea.findByIdAndUpdate(
       req.params.id,
       {
         $set: {
@@ -54,6 +57,10 @@ router.put("/:id", async (req, res) => {
       }
     );
     res.json({ success: true, data: updatedIdea });
+    }
+    //usernames do not match
+    return res.status(403).json({ success: false, error: "you are not authorised to update this resource" })
+    
   } catch (error) {
     console.log(error);
     res.status(500).json({ success: false, error: "something went wrong" });
@@ -62,8 +69,14 @@ router.put("/:id", async (req, res) => {
 // delete  idea
 router.delete("/:id", async (req, res) => {
   try {
-    await Idea.findByIdAndDelete(req.params.id);
-    res.json({ success: true, data: {} });
+    const idea = await Idea.findById(req.params.id)
+    //match the username
+    if(idea.username === req.body.username){
+      await Idea.findByIdAndDelete(req.params.id);
+      return res.json({ success: true, data: {} });
+    }
+    //usernames do not match
+    res.status(403).json({ success: false, error: "your are not authorized to delete this resource" })
   } catch (error) {
     console.log(error);
     res.status(500).json({ success: false, error: "something went wrong" });
